auth code colocation
