+++ src/pages/api/auth/[...nextauth].ts
@@ -4,8 +4,7 @@
  */
 
 import createFetch from '@vercel/fetch'
-import NextAuth, { Account, Session, Profile } from 'next-auth'
-import { JWT } from 'next-auth/jwt'
+import NextAuth from 'next-auth'
 import { URL } from 'url'
 import { AuthErrors } from 'types/auth'
 import CloudIdpProvider from 'lib/auth/cloud-idp-provider'
@@ -27,7 +26,7 @@ export default NextAuth({
 		 * ourselves in this signOut event.
 		 * https://github.com/nextauthjs/next-auth/discussions/3938
 		 */
-		async signOut({ token }: { token: JWT }) {
+		async signOut({ token }) {
 			if (isDev) {
 				console.log('Inside of NextAuth.events.signOut')
 			}
@@ -63,15 +62,7 @@ export default NextAuth({
 		 *
 		 * ref: https://next-auth.js.org/configuration/callbacks#jwt-callback
 		 */
-		async jwt({
-			token,
-			account,
-			profile,
-		}: {
-			token: JWT
-			account?: Account
-			profile?: Profile
-		}) {
+		async jwt({ token, account, profile }) {
 			const isInitial = !!account && !!profile
 			isDev &&
 				console.log('jwt callback (%s)', isInitial ? 'initial' : 'subsequent')
@@ -133,13 +124,7 @@ export default NextAuth({
 		 *
 		 * ref: https://next-auth.js.org/configuration/callbacks#session-callback
 		 */
-		async session({
-			session,
-			token,
-		}: {
-			token: JWT
-			session: Session
-		}): Promise<Session> {
+		async session({ session, token }) {
 			return {
 				...session,
 				accessToken: token.access_token,
